refactor: organize utility headers into utility subfolder
- Move utility headers to include/gecko/core/utility/
  (bit.h, hash.h, random.h, thread.h, time.h)
- Update all include paths to use new utility/ structure
- Fix relative includes in utility headers to use full gecko/ paths
- Final structure: root (infrastructure), services/, utility/
