+++ examples/core_example/src/main.cpp
@@ -1,13 +1,13 @@
 #include "gecko/core/boot.h"
-#include "gecko/core/events.h"
-#include "gecko/core/log.h"
-#include "gecko/core/memory.h"
-#include "gecko/core/modules.h"
-#include "gecko/core/profiler.h"
-#include "gecko/core/random.h"
+#include "gecko/core/services/events.h"
+#include "gecko/core/services/log.h"
+#include "gecko/core/services/memory.h"
+#include "gecko/core/services/modules.h"
+#include "gecko/core/services/profiler.h"
+#include "gecko/core/utility/random.h"
 #include "gecko/core/services.h"
-#include "gecko/core/thread.h"
-#include "gecko/core/time.h"
+#include "gecko/core/utility/thread.h"
+#include "gecko/core/utility/time.h"
 #include "gecko/core/version.h"
 #include "gecko/runtime/console_log_sink.h"
 #include "gecko/runtime/event_bus.h"
