+++ include/gecko/core/services/memory.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "api.h"
-#include "assert.h"
-#include "labels.h"
-#include "types.h"
+#include "gecko/core/api.h"
+#include "gecko/core/assert.h"
+#include "gecko/core/labels.h"
+#include "gecko/core/types.h"
 
 namespace gecko {
