+++ include/gecko/core/utility/bit.h
