+++ include/gecko/core/utility/hash.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "types.h"
+#include "gecko/core/types.h"
 
 namespace gecko {
