+++ include/gecko/runtime/event_bus.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "gecko/core/events.h"
+#include "gecko/core/services/events.h"
 
 #include <atomic>
 #include <deque>
