+++ include/gecko/runtime/ring_profiler.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "gecko/core/jobs.h"
-#include "gecko/core/profiler.h"
+#include "gecko/core/services/jobs.h"
+#include "gecko/core/services/profiler.h"
 
 #include <atomic>
 #include <mutex>
