+++ include/gecko/runtime/thread_pool_job_system.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "gecko/core/jobs.h"
+#include "gecko/core/services/jobs.h"
 
 #include <atomic>
 #include <condition_variable>
