+++ src/core/events.cpp
@@ -1,10 +1,10 @@
-#include "gecko/core/events.h"
+#include "gecko/core/services/events.h"
 
 #include "gecko/core/assert.h"
 #include "gecko/core/labels.h"
-#include "gecko/core/log.h"
-#include "gecko/core/profiler.h"
-#include "labels.h"
+#include "gecko/core/services/log.h"
+#include "gecko/core/services/profiler.h"
+#include "private/labels.h"
 
 namespace gecko {
