+++ src/core/random.cpp
@@ -1,9 +1,9 @@
-#include "gecko/core/random.h"
+#include "gecko/core/utility/random.h"
 
 #include "gecko/core/assert.h"
-#include "gecko/core/log.h"
-#include "gecko/core/profiler.h"
-#include "labels.h"
+#include "gecko/core/services/log.h"
+#include "gecko/core/services/profiler.h"
+#include "private/labels.h"
 
 #include <cstring>
 #include <random>
