+++ src/core/services.cpp
@@ -1,9 +1,9 @@
 #include "gecko/core/services.h"
 
 #include "gecko/core/assert.h"
-#include "gecko/core/log.h"
-#include "gecko/core/profiler.h"
-#include "labels.h"
+#include "gecko/core/services/log.h"
+#include "gecko/core/services/profiler.h"
+#include "private/labels.h"
 
 #include <atomic>
 #include <cstddef>
