+++ src/runtime/ring_logger.cpp
@@ -1,12 +1,12 @@
 #include "gecko/runtime/ring_logger.h"
 
 #include "gecko/core/assert.h"
-#include "gecko/core/jobs.h"
-#include "gecko/core/log.h"
-#include "gecko/core/profiler.h"
-#include "gecko/core/thread.h"
-#include "gecko/core/time.h"
-#include "labels.h"
+#include "gecko/core/services/jobs.h"
+#include "gecko/core/services/log.h"
+#include "gecko/core/services/profiler.h"
+#include "gecko/core/utility/thread.h"
+#include "gecko/core/utility/time.h"
+#include "private/labels.h"
 
 #include <atomic>
 #include <cstdarg>
