+++ src/runtime/ring_profiler.cpp
@@ -1,10 +1,10 @@
 #include "gecko/runtime/ring_profiler.h"
 
 #include "gecko/core/assert.h"
-#include "gecko/core/bit.h"
-#include "gecko/core/thread.h"
-#include "gecko/core/time.h"
-#include "labels.h"
+#include "gecko/core/utility/bit.h"
+#include "gecko/core/utility/thread.h"
+#include "gecko/core/utility/time.h"
+#include "private/labels.h"
 
 #include <algorithm>
 #include <vector>
