+++ src/runtime/tracking_allocator.cpp
@@ -1,9 +1,9 @@
 #include "gecko/runtime/tracking_allocator.h"
 
 #include "gecko/core/assert.h"
-#include "gecko/core/log.h"
-#include "gecko/core/profiler.h"
-#include "labels.h"
+#include "gecko/core/services/log.h"
+#include "gecko/core/services/profiler.h"
+#include "private/labels.h"
 
 #include <atomic>
 #include <mutex>
